ExitStatus::signal_name with the name of the signal
It's useful to know the actual signal name corresponding to the signal number
(e.g. 11 = "Segmentation Fault"). This commit does not alter the API, it just
adds a method which returns Some(String) when the ExitStatus is actually a
Signal, otherwise it returns None.

The signal name is available only under unix systems and it's obtained with
`strsignal`. The unsafe block is due to the FFI call. When the signal name is
not available (i.e. not unix or unknown signal number), `None` is returned
instead.