Make sure the child process exits also on errors
When the child process failes to exec() we need to make sure that the
fork exits, and does not run the parent code. This caused a quite nasty
bug during the sandbox cleanup:

- The child process sets up the various mountpoints in the user
  namespace
- There is an error and the exec() is not called
- The child does not exit() so the destructors started running *also* in
  the child process, including the one of the tmpdir of the sandbox
- The tmpdir Drop starts removing the sandbox content *including* the
  content of the bind mounts, removing also the files outside the
  sandbox

When the destructors are called by the parent process the mountpoint
contents are not removed since those were user-namespace mounts and are
not visible by the parent.

The solution is to make sure those destructors will never run, and by
using std::process:exit() [1] we make sure of that.

[1]: https://doc.rust-lang.org/std/process/fn.exit.html