linux/seccomp_filter: accept pseudo syscall numbers
If the given architecture does not have the given system call,
then a negative pseudo system call number is returned. This is
not an error, and can be passed to other libseccomp functions.

In the case of an unknown syscall name the constant
`__NR_SCMP_ERROR` is returned, so check for that.

For example, aarch64 does not have a separate `chmod()` system call,
so task-maker-rust was not really usable because pseudo system call
numbers were rejected.